+++ CHANGES.rst
@@ -5,6 +5,8 @@ Changelog
 Unreleased
 ----------
 
+* Include ``sys.node_checks`` for checking cluster healthiness.
+
 2.22.0 (2023-01-31)
 -------------------